refactor: :recycle: move theme code out of common-defer
